Add base app including OTP services
This commit adds the base app to the repository, including some
already-implemented OTP services for standard HOTP and TOTP and for
whatever standard Valve thinks it's using for Steam.

The functionality of the SteamTotpService has not yet been verified, but
its implementation is based on the implementation of WinAuth's
Steam-specific code (credit where credit is due).

The app itself does not yet have any kind of functionality, it's just the
default Android Studio template you get when creating a new Android app.